+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file. Take a look
 #### Navigator
 
 * Fixed memory leaks in the EPUB and PDF navigators.
+* [#61](https://github.com/readium/swift-toolkit/issues/61) Fixed serving EPUB resources when the HREF contains an anchor or query parameters.
 
 #### Streamer